backend/pyproject.toml
[project]
name = "askdata-backend"
version = "0.1.0"
description = "Backend service for AskData."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.116.0,<1.0.0",
  "httpx>=0.28.0,<1.0.0",
  "openai>=1.107.0,<2.0.0",
  "pydantic-settings>=2.10.0,<3.0.0",
  "psycopg[binary]>=3.2.0,<4.0.0",
  "pytest>=8.4.0,<9.0.0",
  "sqlglot>=27.13.2,<28.0.0",
  "uvicorn[standard]>=0.35.0,<1.0.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
